pyproject.toml
[project]
name = "aiframeworks"
version = "0.3.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "imageio==2.37.0",
    "joblib==1.5.2",
    "lazy-loader==0.4",
    "networkx==3.5",
    "numpy==2.2.6",
    "onnxruntime>=1.22.1",
    "opencv-python==4.12.0.88",
    "packaging==25.0",
    "pillow==11.3.0",
    "scikit-image==0.25.2",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "threadpoolctl==3.6.0",
    "tifffile==2025.9.9",
    "torch>=2.8.0",
    "uniface>=0.1.8",
]
